+++ lib/api/src/grpc/conversions.rs
@@ -353,8 +353,8 @@ impl From<segment::data_types::vectors::VectorType> for Vector {
     }
 }
 
-impl From<segment::data_types::vectors::NamedVectors> for NamedVectors {
-    fn from(vectors: segment::data_types::vectors::NamedVectors) -> Self {
+impl From<segment::data_types::named_vectors::NamedVectors> for NamedVectors {
+    fn from(vectors: segment::data_types::named_vectors::NamedVectors) -> Self {
         Self {
             vectors: vectors
                 .into_iter()
@@ -398,7 +398,7 @@ impl From<segment::types::ScoredPoint> for ScoredPoint {
     }
 }
 
-impl From<NamedVectors> for segment::data_types::vectors::NamedVectors {
+impl From<NamedVectors> for segment::data_types::named_vectors::NamedVectors {
     fn from(vectors: NamedVectors) -> Self {
         vectors
             .vectors
