+++ lib/collection/src/collection_manager/segments_searcher.rs
@@ -4,7 +4,8 @@ use std::sync::Arc;
 use futures::future::try_join_all;
 use itertools::Itertools;
 use parking_lot::RwLock;
-use segment::data_types::vectors::{NamedVectors, VectorElementType};
+use segment::data_types::named_vectors::NamedVectors;
+use segment::data_types::vectors::VectorElementType;
 use segment::entry::entry_point::OperationError;
 use segment::spaces::tools::peek_top_largest_scores_iterable;
 use segment::types::{
