+++ lib/segment/src/data_types/mod.rs
@@ -1,2 +1,3 @@
 pub mod text_index;
 pub mod vectors;
+pub mod named_vectors;
