+++ lib/segment/src/data_types/named_vectors.rs
@@ -1,0 +1,4 @@
+use std::collections::HashMap;
+use super::vectors::VectorElementType;
+
+pub type NamedVectors = HashMap<String, Vec<VectorElementType>>;
