+++ lib/segment/src/segment.rs
@@ -13,7 +13,8 @@ use tar::Builder;
 use crate::common::file_operations::{atomic_save_json, read_json};
 use crate::common::version::StorageVersion;
 use crate::common::{check_vector_name, check_vectors_set};
-use crate::data_types::vectors::{NamedVectors, VectorElementType};
+use crate::data_types::named_vectors::NamedVectors;
+use crate::data_types::vectors::VectorElementType;
 use crate::entry::entry_point::OperationError::ServiceError;
 use crate::entry::entry_point::{
     get_service_error, OperationError, OperationResult, SegmentEntry, SegmentFailedState,
